Move async, sync, dataflow and apply to executors
