+++ libs/async/include/hpx/async/sync.hpp
@@ -13,7 +13,7 @@
 #include <hpx/async/sync.hpp>
 #include <hpx/async/sync_fwd.hpp>
 #include <hpx/execution/detail/sync_launch_policy_dispatch.hpp>
-#include <hpx/execution/sync.hpp>
+#include <hpx/executors/sync.hpp>
 #include <hpx/functional/bind_back.hpp>
 #include <hpx/functional/traits/is_action.hpp>
 #include <hpx/local_async/sync.hpp>
