+++ libs/executors/include/hpx/executors/guided_pool_executor.hpp
@@ -15,7 +15,7 @@
 
 #include <hpx/debugging/demangle_helper.hpp>
 #include <hpx/debugging/print.hpp>
-#include <hpx/execution/dataflow.hpp>
+#include <hpx/executors/dataflow.hpp>
 #include <hpx/executors/pool_executor.hpp>
 #include <hpx/functional/bind_back.hpp>
 #include <hpx/functional/invoke.hpp>
