+++ libs/local_async/include/hpx/local_async/apply.hpp
@@ -9,6 +9,6 @@
 
 #include <hpx/config.hpp>
 #include <hpx/async_base/apply.hpp>
-#include <hpx/execution/apply.hpp>
+#include <hpx/executors/apply.hpp>
 
 #endif
