+++ libs/local_async/include/hpx/local_async/async.hpp
@@ -9,6 +9,6 @@
 
 #include <hpx/config.hpp>
 #include <hpx/async_base/async.hpp>
-#include <hpx/execution/async.hpp>
+#include <hpx/executors/async.hpp>
 
 #endif
